Correct several exporter uses of file_basepath with a dedupe
